+++ src/components/Business/BusinessesList.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/variables";
+@import "../../styles/variables.scss";
 
 .businesses {
   padding-top: 3rem;
