+++ src/components/Categories/CategoriesList.jsx
@@ -1,6 +1,6 @@
 import styles from "./CategoriesList.module.scss";
-import { SingleCategory } from "./SingleCategory";
-import { categories } from "../data/data";
+import { SingleCategory } from "../Categories/SingleCategory";
+import { categories } from "../../data/data";
 
 export function CategoriesList() {
   return (
