+++ src/components/Categories/CategoriesList.module.scss
