+++ src/components/InputSearch/index.js
@@ -1,0 +1,1 @@
+export * from "./InputSearch";
