+++ src/components/LoginButton/LoginButton.jsx
@@ -1,6 +1,6 @@
 import styles from "./LoginButton.module.scss";
 import { useNavigate } from "react-router-dom";
-import { ROUTES } from "../router/consts";
+import { ROUTES } from "../../router/consts";
 
 export function LoginButton() {
   const navigate = useNavigate();
