+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
-import { Hero } from "../components/Hero";
-import { InputSearch } from "../components/InputSearch";
-import { CategoriesList } from "../components/CategoriesList";
-import { BusinessesList } from "../components/BusinessesList";
+import { Hero } from "../components/Hero/Hero";
+import { InputSearch } from "../components/InputSearch/InputSearch";
+import { CategoriesList } from "../components/Categories/CategoriesList";
+import { BusinessesList } from "../components/Business/BusinessesList";
 
 export function Home() {
   return (
