+++ vite.config.js
@@ -7,6 +7,7 @@ export default defineConfig({
   css: {
     preprocessorOptions: {
       scss: {
+        silenceDeprecations: ["legacy-js-api"],
         additionalData: `@import "./src/styles/variables";`
       }
     }
